requirements.txt
# Core dependencies
pyTelegramBotAPI>=4.12.0
openai>=1.0.0
python-dotenv>=1.0.0

# Document processing
PyPDF2>=3.0.0
python-docx>=0.8.11
python-pptx>=0.6.21
openpyxl>=3.1.2

# Text processing and utilities
requests>=2.28.0
beautifulsoup4>=4.11.0
chardet>=5.2.0

# Optional: Alternative AI providers
# google-generativeai>=0.3.0  # Uncomment if using Gemini

# Development dependencies (optional)
# pytest>=7.0.0
# pytest-cov>=4.0.0
# flake8>=6.0.0
